# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "api 计算Bestdori谱面难度，获得Bestdori数据，常见Bandori谱面格式转换等",
        "title": "ayachan API",
        "contact": {
            "name": "6QHTSK"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://mit-license.org/"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/v2",
    "paths": {
        "/chart-display": {
            "get": {
                "description": "获取谱面展示数据库的谱面",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chartDisplay"
                ],
                "summary": "获取谱面展示数据库的谱面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ChartDisplayOutput"
                        }
                    }
                }
            }
        },
        "/chart-display/:chartID": {
            "get": {
                "description": "获取谱面展示数据库的指定ID的谱面",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chartDisplay"
                ],
                "summary": "获取谱面展示数据库的谱面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "谱面ID",
                        "name": "chartID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ChartDisplayOutputID"
                        }
                    }
                }
            }
        },
        "/charter-like-rank": {
            "get": {
                "description": "获取谱师获赞数的排行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师获赞数的排行",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
...
        "/charter-post-rank": {
            "get": {
                "description": "获取谱师发表谱面的排行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师发表谱面的排行",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
        "/charter/:charter/basic-info": {
            "get": {
                "description": "获取谱师的基本信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师的基本信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "谱师用户名",
                        "name": "charter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "谱师基础信息",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
        "/charter/:charter/like-rank": {
            "get": {
                "description": "获取谱师发表的赞数排行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师发表的赞数排行",
                "parameters": [
                    {
                        "type": "string",
                        "description": "谱师用户名",
                        "name": "charter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "谱面列表",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
        "/charter/:charter/note-rank": {
            "get": {
                "description": "获取谱师发表的音符数排行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师发表的音符数排行",
                "parameters": [
                    {
                        "type": "string",
                        "description": "谱师用户名",
                        "name": "charter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "谱面列表",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
        "/charter/:charter/nps-rank": {
            "get": {
                "description": "获取谱师发表的谱面的NPS排行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师发表的谱面的NPS排行",
                "parameters": [
                    {
                        "type": "string",
                        "description": "谱师用户名",
                        "name": "charter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "谱面列表",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
        "/charter/:charter/post": {
            "get": {
                "description": "获取谱师发表的谱面",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师发表的谱面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "谱师用户名",
                        "name": "charter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "谱面列表",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
        "/charter/:charter/time-rank": {
            "get": {
                "description": "获取谱师发表的谱面的时间排行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱师发表的谱面的时间排行",
                "parameters": [
                    {
                        "type": "string",
                        "description": "谱师用户名",
                        "name": "charter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "谱面列表",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
...
        "/song-like-rank": {
            "get": {
                "description": "获取谱面获赞数的排行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BestdoriInfo"
                ],
                "summary": "获取谱面获赞数的排行",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码,默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面限制，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/Controllers.InfoOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Controllers.ChartDisplayOutput": {
            "type": "object",
            "properties": {
                "chart_set": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chartFormat.Chart"
                    }
                },
                "result": {
                    "type": "boolean"
                }
            }
        },
        "Controllers.ChartDisplayOutputID": {
            "type": "object",
            "properties": {
                "chart_set": {
                    "$ref": "#/definitions/chartFormat.Chart"
                },
                "result": {
                    "type": "boolean"
                }
            }
        },
        "Controllers.InfoOutput": {
            "type": "object",
            "properties": {
                "list": {},
                "result": {
                    "type": "boolean"
                }
            }
        },
...
        "chartFormat.Chart": {
            "type": "object",
            "properties": {
                "artists": {
                    "description": "谱面的艺术家",
                    "type": "string"
                },
                "author": {
                    "type": "object",
                    "properties": {
                        "authorID": {
                            "description": "谱面作者用户名",
                            "type": "integer"
                        },
                        "authorName": {
                            "description": "Author_Other 其他作者时使用的名义",
                            "type": "string"
                        }
                    }
                },
                "chartID": {
                    "description": "已提交：BanGround社区谱面ID,未提交：Bestdori社区Ex/SpID",
                    "type": "integer"
                },
                "diffs": {
                    "type": "object",
                    "properties": {
                        "bestdoriChartID": {
                            "description": "Bestdori的谱面ID 0 就是未发表",
                            "type": "integer"
                        },
                        "bpmhigh": {
                            "type": "number"
                        },
                        "bpmlow": {
                            "type": "number"
                        },
                        "diff": {
                            "type": "integer"
                        },
                        "irregular": {
                            "description": "存在多压/交叉（出张）0 失败 1 标准 2 非标准",
                            "type": "integer"
                        },
                        "irregularInfo": {
                            "description": "无法分析的第一个错误情况",
                            "type": "string"
                        },
                        "level": {
                            "type": "integer"
                        },
                        "mainBPM": {
                            "type": "number"
                        },
                        "totalNPS": {
                            "type": "number"
                        },
                        "totalNote": {
                            "type": "integer"
                        },
                        "totalTime": {
                            "type": "number"
                        }
                    }
                },
                "rel": {
                    "description": "访问网址",
                    "type": "string"
                },
                "songUrl": {
                    "description": "谱面资源",
                    "type": "object",
                    "properties": {
                        "audio": {
                            "type": "string"
                        },
                        "cover": {
                            "type": "string"
                        }
                    }
                },
                "submitted": {
                    "description": "是否已经提交到BanGround社区",
                    "type": "boolean"
                },
                "title": {
                    "description": "谱面的标题",
                    "type": "string"
                }
            }
        },
...
    }
}